day 30

+++ handlers/refill.py
@@ -1,5 +1,6 @@
 import asyncio
-import re
+import random
+
 import decouple
 import shutup
 
@@ -7,10 +8,10 @@ from aiogram import Dispatcher, types
 from aiogram.dispatcher import FSMContext
 from aiogram.dispatcher.filters.state import StatesGroup, State
 
-
 from keyboards import inline
-from database import users, documents, binance_db, balance
+from database import users, documents, binance_db, balance, thedex_db
 from binance import actions as binance
+from handlers.refill_500 import registration_500
 
 shutup.please()
 
@@ -48,44 +49,47 @@ async def refill_handler(call: types.CallbackQuery):
     language = await users.user_data(call.from_user.id)
     status = await documents.status_docs(call.from_user.id)
     docs_status = await balance.get_document_status(call.from_user.id)
+    trans = await thedex_db.get_transaction(call.from_user.id)
     try:
         docs_status = docs_status[0]
     except TypeError:
         docs_status = None
-    if docs_status:
+    if not trans:
+        if docs_status:
+            if not status:
 
-        if not status:
-
-            photo = decouple.config("BANNER_REFILL")
-            text = 'Условия участия зависят от суммы размещенных криптоактивов. Мы рекомендуем изучить подробную ' \
-                   'информацию  о каждом варианте, до пополнения баланса. ' \
-                   'Если Вы уже знаете все условия, то можете переходить к пополнению.'
-            if language[4] == "EN":
-                text = "The terms of participation depend on the amount of crypto assets you have deposited. " \
-                       "We recommend reviewing detailed information about each option before depositing funds. " \
-                       "If you are already familiar with all the terms, you can proceed with the deposit."
-            try:
-                await call.message.delete()
-            except:
-                pass
-            await call.message.answer_photo(photo, text, reply_markup=inline.refill_main_menu(language[4]))
-        else:
-
-            contract = await documents.check_approve_contract(call.from_user.id)
-
-            if contract is False:
-                text = 'Документы пока еще не подтверждены, ожидайте'
+                photo = decouple.config("BANNER_REFILL")
+                text = 'Условия участия зависят от суммы размещенных криптоактивов. Мы рекомендуем изучить подробную ' \
+                       'информацию  о каждом варианте, до пополнения баланса. ' \
+                       'Если Вы уже знаете все условия, то можете переходить к пополнению.'
                 if language[4] == "EN":
-                    text = "The documents are not yet confirmed. Please wait."
-                await call.message.delete()
-                await call.message.answer(text)
+                    text = "The terms of participation depend on the amount of crypto assets you have deposited. " \
+                           "We recommend reviewing detailed information about each option before depositing funds. " \
+                           "If you are already familiar with all the terms, you can proceed with the deposit."
+                try:
+                    await call.message.delete()
+                except:
+                    pass
+                await call.message.answer_photo(photo, text, reply_markup=inline.refill_main_menu(language[4]))
             else:
-                await binanceapi_step1_call(call)
 
+                contract = await documents.check_approve_contract(call.from_user.id)
+
+                if contract is False:
+                    text = 'Документы пока еще не подтверждены, ожидайте'
+                    if language[4] == "EN":
+                        text = "The documents are not yet confirmed. Please wait."
+                    await call.message.delete()
+                    await call.message.answer(text)
+                else:
+                    await binanceapi_step1_call(call)
+
+        else:
+            await call.message.delete()
+            await new_docs(call)
     else:
+        await registration_500(call)
 
-        await call.message.delete()
-        await new_docs(call)
 
 
 async def new_docs(call: types.CallbackQuery):
@@ -323,10 +327,10 @@ async def biguser_registration(call: types.CallbackQuery):
 async def biguser_registration_step_1(call: types.CallbackQuery, state: FSMContext):
     language = await users.user_data(call.from_user.id)
     if call.data == 'no':
-        text = 'Чтобы воспользоваться данным предложением Вам необходимо зарегистрировать аккаунт ' \
-               'на бирже Binance и пройти KYC верификацию, ' \
-               'после этого вы сможете продолжить процедуру регистрации в программе ' \
-               'управляемых субаккаунтов.'
+        text = 'Чтобы воспользоваться данным форматом пополнения, ' \
+               'Вам необходимо зарегистрировать аккаунт на бирже Binance, ' \
+               'а также пройти KYC верификацию Plus, с подтверждением адреса.' \
+               '\n\nПосле этого вы сможете продолжить процедуру регистрации в программе управляемых субаккаунтов.'
         if language[4] == "EN":
             text = "To take advantage of this offer, you need to register an account on the Binance exchange " \
                    "and complete the KYC verification process. After that, you can proceed with the registration " \
@@ -726,7 +730,7 @@ async def count_refill(msg: types.Message, state: FSMContext):
 
 
 def register(dp: Dispatcher):
-    dp.register_callback_query_handler(refill_handler, text='refill')
+    dp.register_callback_query_handler(refill_handler, text='refill', state="*")
     dp.register_callback_query_handler(handle_deposit_funds, text='deposit_funds')
     dp.register_callback_query_handler(handle_review_terms, text='review_terms')
     dp.register_callback_query_handler(handle_distribution, text='distribution')
@@ -739,9 +743,9 @@ def register(dp: Dispatcher):
     # dp.register_message_handler(biguser_registration_step3, content_types=['text', 'video', 'photo', 'document'],
     #                             state=BigUser.contract)
     dp.register_message_handler(binanceapi_step1_msg, state=BigUser.finish)
-    dp.register_message_handler(binanceapi_step2, state=BinanceAPI.alias)
-    dp.register_message_handler(binance_step3, state=BinanceAPI.api_key)
-    dp.register_message_handler(binance_step4, state=BinanceAPI.api_secret)
+    # dp.register_message_handler(binanceapi_step2, state=BinanceAPI.alias)
+    # dp.register_message_handler(binance_step3, state=BinanceAPI.api_key)
+    # dp.register_message_handler(binance_step4, state=BinanceAPI.api_secret)
     dp.register_message_handler(count_refill, state=Refill.count)
     dp.register_callback_query_handler(new_docs_2, state=NewDoc.docs)
     dp.register_callback_query_handler(new_docs_3, state=NewDoc.docs_2)
